+++ src/main/java/com/justbelieveinmyself/basedknowlage/entity/User.java
@@ -1,6 +1,6 @@
 package com.justbelieveinmyself.basedknowlage.entity;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +17,7 @@ import java.io.Serializable;
 public class User implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
     private String name;
     @Column(name = "pass")
     private String password;
