+++ src/main/java/com/justbelieveinmyself/manytomany/entity/Laptop3.java
@@ -1,7 +1,6 @@
 package com.justbelieveinmyself.manytomany.entity;
 
-import com.justbelieveinmyself.onetoone.entity.Laptop;
-import jakarta.persistence.*;
+import javax.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,11 +17,11 @@ import java.util.List;
 @Setter
 public class Laptop3 {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long lid;
+    private int lid;
     private String name;
     @ManyToMany
     private List<Student3> student3 = new ArrayList<>();
-    public Laptop3(long lid, String name, List<Student3> student3){
+    public Laptop3(int lid, String name, List<Student3> student3){
         this.lid = lid;
         this.name = name;
         this.student3 = student3;
